+++ packages/flutter_timeline/lib/src/flutter_timeline_navigator_userstory.dart
@@ -92,7 +92,7 @@ Widget _timelineScreenRoute({
         appBar: AppBar(
           backgroundColor: const Color(0xff212121),
           title: Text(
-            config.optionsBuilder(context).translations.timeLineScreenTitle!,
+            config.optionsBuilder(context).translations.timeLineScreenTitle,
             style: TextStyle(
               color: Theme.of(context).primaryColor,
               fontSize: 24,
@@ -246,7 +246,7 @@ Widget _postCreationScreenRoute({
           backgroundColor: const Color(0xff212121),
           leading: backButton,
           title: Text(
-            config.optionsBuilder(context).translations.postCreation!,
+            config.optionsBuilder(context).translations.postCreation,
             style: TextStyle(
               color: Theme.of(context).primaryColor,
               fontSize: 24,
@@ -313,7 +313,7 @@ Widget _postOverviewScreenRoute({
           leading: backButton,
           backgroundColor: const Color(0xff212121),
           title: Text(
-            config.optionsBuilder(context).translations.postOverview!,
+            config.optionsBuilder(context).translations.postOverview,
             style: TextStyle(
               color: Theme.of(context).primaryColor,
               fontSize: 24,
@@ -372,7 +372,7 @@ Widget _postCategorySelectionScreen({
           leading: backButton,
           backgroundColor: const Color(0xff212121),
           title: Text(
-            config.optionsBuilder(context).translations.postCreation!,
+            config.optionsBuilder(context).translations.postCreation,
             style: TextStyle(
               color: Theme.of(context).primaryColor,
               fontSize: 24,