+++ packages/flutter_timeline_view/lib/src/widgets/reaction_bottom.dart
@@ -65,7 +65,7 @@ class _ReactionBottomState extends State<ReactionBottom> {
                   ],
                 ),
               ),
-              widget.translations.writeComment!,
+              widget.translations.writeComment,
             ),
           ),
         ),